benchmarks/speed/requirements.txt
transformers>=3.0.0,<3.1.0
spacy-transformers>=1.0.0a23,<1.1.0
cupy-cuda102
stanza>=1.0.0,<1.2.0
flair>=0.6.0,<1.0.0
ufal.udpipe
